pyproject.toml
[tool.poetry]
name = "youtube-seo-tool"
version = "0.1.0"
description = "Integrates a multi-LLM chat interface with a YouTube SEO generator"
authors = ["SlySlayer32 <slyslayerfx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requires-python = ">=3.13"
Flask = "^3.1.0"
requests = "^2.32.3"
google-api-python-client = "^2.120.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^1.1.0"
ffmpeg-python = "^0.2.0"
tenacity = "^8.2.3"
redis = "^5.0.1"
prometheus_client = "^0.23.0"
Flask-Cors = "^4.0.0"
SQLAlchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
transformers = "^4.35.2"
torch = "^2.1.1"
sentencepiece = "^0.1.99"
tiktoken = "^0.5.1"
langchain = "^0.0.330"
faiss-cpu = "^1.7.4"
openai = "^1.3.7"
chromadb = "^0.4.16"
sentence-transformers = "^2.2.2"
Flask-Migrate = "^5.0.0"
Flask-SQLAlchemy = "^3.1.1"
Flask-WTF = "^1.2.0"
Flask-Login = "^0.6.2"
Flask-Babel = "^3.0.0"
Flask-Bootstrap = "^5.3.0"
email-validator = "^2.1.0"
werkzeug = "^3.0.1"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.10.0"
requests-mock = "^1.12.1"
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
exclude = '/(\.git|\.hg|\.mypy_cache|\.tox|\.venv|_build|build|dist)/'

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
exclude = ".git,__pycache__,build,dist"
max-line-length = 100
ignore = "E203, W503"
